+++ src/common/librouter/Makefile.am
@@ -65,7 +65,8 @@ test_ldadd = \
         $(top_builddir)/src/common/libzmqutil/libzmqutil.la \
         $(top_builddir)/src/common/libflux-internal.la \
         $(top_builddir)/src/common/libflux-core.la \
-        $(top_builddir)/src/common/libtap/libtap.la
+        $(top_builddir)/src/common/libtap/libtap.la \
+        $(ZMQ_LIBS)
 
 test_cppflags = \
         $(AM_CPPFLAGS) \
