+++ src/common/libterminus/Makefile.am
@@ -39,7 +39,8 @@ test_ldadd = \
 	$(top_builddir)/src/common/libzmqutil/libzmqutil.la \
         $(top_builddir)/src/common/libflux-internal.la \
         $(top_builddir)/src/common/libflux-core.la \
-        $(top_builddir)/src/common/libtap/libtap.la
+        $(top_builddir)/src/common/libtap/libtap.la \
+        $(ZMQ_LIBS)
 
 test_ldflags = \
 	-no-install
