+++ src/common/libzmqutil/Makefile.am
@@ -18,10 +18,13 @@ noinst_LTLIBRARIES = \
 libzmqutil_la_SOURCES = \
 	msg_zsock.h \
 	msg_zsock.c \
+	reactor.h \
+	reactor.c \
 	ev_zmq.h \
 	ev_zmq.c
 
 TESTS = test_msg_zsock.t \
+	test_reactor.t \
 	test_ev.t
 
 check_PROGRAMS = \
@@ -49,6 +52,11 @@ test_msg_zsock_t_SOURCES = test/msg_zsock.c
 test_msg_zsock_t_CPPFLAGS = $(test_cppflags)
 test_msg_zsock_t_LDADD = $(test_ldadd)
 test_msg_zsock_t_LDFLAGS = $(test_ldflags)
+
+test_reactor_t_SOURCES = test/reactor.c
+test_reactor_t_CPPFLAGS = $(test_cppflags)
+test_reactor_t_LDADD = $(test_ldadd)
+test_reactor_t_LDFLAGS = $(test_ldflags)
 
 test_ev_t_SOURCES = test/ev.c
 test_ev_t_CPPFLAGS = $(test_cppflags)
